requirements.txt
dnspython==2.7.0
numpy==2.2.6
pandas==2.3.0
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
